+++ ui-test/src/test/java/kz/yandex/scooter/OrderedRunner.java
@@ -1,0 +1,32 @@
+package kz.yandex.scooter;
+
+import org.junit.runners.BlockJUnit4ClassRunner;
+import org.junit.runners.model.FrameworkMethod;
+import org.junit.runners.model.InitializationError;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class OrderedRunner extends BlockJUnit4ClassRunner {
+    public OrderedRunner(Class<?> clazz) throws InitializationError {
+        super(clazz);
+    }
+
+    @Override
+    protected List<FrameworkMethod> computeTestMethods() {
+        List<FrameworkMethod> list = super.computeTestMethods();
+        List<FrameworkMethod> copy = new ArrayList<>(list);
+        Collections.sort(copy, (f1, f2) -> {
+            TestOrder o1 = f1.getAnnotation(TestOrder.class);
+            TestOrder o2 = f2.getAnnotation(TestOrder.class);
+
+            if (o1 == null || o2 == null) {
+                return -1;
+            }
+
+            return o1.value() - o2.value();
+        });
+        return copy;
+    }
+}
