+++ ui-test/src/test/java/kz/yandex/scooter/TestOrder.java
@@ -1,0 +1,12 @@
+package kz.yandex.scooter;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+public @interface TestOrder {
+    int value() default 0;
+}
