--- ui-test/src/test/java/yandex_scooter/ui_test/TestOrder.java
@@ -1,12 +1,0 @@
-package yandex_scooter.ui_test;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface TestOrder {
-    int value() default 0;
-}
